split types into distinct files
