+++ beacon-chain/state/state-native/custom-types/BUILD.bazel
@@ -2,7 +2,12 @@ load("@prysm//tools/go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["state_arrays.go"],
+    srcs = [
+        "block_roots.go",
+        "byte32.go",
+        "randao_mixes.go",
+        "state_roots.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state/state-native/custom-types",
     visibility = ["//visibility:public"],
     deps = [
@@ -13,7 +18,12 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["state_arrays_test.go"],
+    srcs = [
+        "block_roots_test.go",
+        "byte32_test.go",
+        "randao_mixes_test.go",
+        "state_roots_test.go",
+    ],
     embed = [":go_default_library"],
     deps = ["//config/fieldparams:go_default_library"],
 )
